requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
openai==1.3.5
python-dotenv==1.0.0
selenium==4.15.2
playwright==1.39.0
fake-useragent==1.3.0
tqdm==4.66.1 